Fix off-heap prototype in OSX
The obvious solution to the memory laziness of OSX is
double mapping. However, we would need to make in-heap
shared and not off-heap. That's because, if we do make
off-heap shared, then we need some way to free the double
mapped regions once we're done with the objects. In this
scenario, the entire in-heap is private with the excpetion
of the double mapped regions. But once we're done with those
regions, we want to map them back to private. So far so good.
When we do that, we still need to free the off-heap region
that was used to double map, because they're still holding
onto the physical memory that was used to double map. And
using madvise() and msync() won't suffice because of how
OSX manages memory (it's lazy). Therefore, we would need
to mmap that off-heap region as well, but since the
off-heap is backed by a file descriptor then wwe cannot
simply mmap those regions or munmap them. Hence, we're
back to the original problem of lazilly releasing
memory back to OS.

With that, the only solution is to make in-heap shared
and associate file descriptor with it. Every time double
mapping happens, we're reusing the physical address of
that mapping that posseses the file descriptor, which in
this case is the in-heap. Since the in-heap will always
be shared and have the file descriptor assciated with it,
we won't need to free it's memory. The only memory range
that needs to be freed is the off-heap regions, and here
since the off-heap is MAP_PRIVATE except for the double
mapped regions, we can just mmap() those regions back to
MAP_PRIVATE.

Signed-off-by: Igor Braga <higorb1@gmail.com>